Change helm chart to helm 3 apiVersion
The chart is already in the format of a Helm 3 chart, however the `apiVersion` of the chart is set to `v1`, which is used for Helm 2 charts. Helm 3 charts should use `apiVersion: v2`.

This makes tools such as ArgoCD mistake the chart for a Helm 2 chart, and therefore template the chart using the wrong helm version. In ArgoCDs case, that means that the CRDs aren't deployed.

reference: https://helm.sh/docs/faq/#chartyaml-apiversion-bump